requirements.txt
setuptools==18.5
google_api_python_client==1.6.1
termcolor==1.1.0
PyYAML==3.12
